Add missing mpl header, move inflows schema to own file
Think this got lost in a rebase?
