+++ flowmachine/flowmachine/core/server/query_schemas/inflows.py
@@ -7,19 +7,19 @@ from marshmallow.validate import OneOf
 
 from flowmachine.core.server.query_schemas.flows import FlowsExposed, FlowsSchema
 from flowmachine.features import Flows
-from flowmachine.features.location.redacted_in_out_flows import RedactedInOutFlow
+from flowmachine.features.location.redacted_flows import RedactedInOutFlow
 
 
-class OutflowsExposed(FlowsExposed):
+class InflowsExposed(FlowsExposed):
     @property
     def _flowmachine_query_obj(self):
         loc1 = self.from_location._flowmachine_query_obj
         loc2 = self.to_location._flowmachine_query_obj
         return RedactedInOutFlow(
-            in_out_flows=Flows(loc1, loc2, join_type=self.join_type).outflow()
+            flows=Flows(loc1, loc2, join_type=self.join_type).inflow()
         )
 
 
-class OutflowsSchema(FlowsSchema):
-    query_kind = fields.String(validate=OneOf(["outflows"]))
-    __model__ = OutflowsExposed
+class InflowsSchema(FlowsSchema):
+    query_kind = fields.String(validate=OneOf(["inflows"]))
+    __model__ = InflowsExposed
